Fix to update peer dimensions when video screen resumes foreground
[Issue/Cause of defect]
If partner device is rotated during video screen of target device is in
background, and then resumes to foreground, partner image will be shown
with incorrect dimensions.

Because when video screen goes to background, VideoCallPresenter stops
checking CVO(Coordination of Video Orientation) of partner device. And
when video screen is resumed to foreground, there is no logic to update
partner image with latest peer dimensions.

[How to fix]
Even if video screen is in background, CVO(peer dimensions) is notified
to DialerCall. Therefore DialerCall should store the dimensions. And
when video screen resumes to fore, VideoCallPresenter should update the
screen with correct(latest) peer dimensions.

Test: manual - Checked that the partner image is shown with correct
dimensions.
Bug: 111575038

Change-Id: I32ff5407f1222b232b47a35e7083a473be67b468
Signed-off-by: ghou8s <ghou8s@gmail.com>